requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Progress Bars
tqdm>=4.65.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
jupyter>=1.0.0
ipython>=8.12.0
notebook>=6.5.4

# Data Validation
jsonschema>=4.17.0

# Accelerated Training (optional)
accelerate>=0.20.0

# Model Optimization (optional)
onnx>=1.14.0
onnxruntime>=1.15.0

# Additional NLP Tools
nltk>=3.8.1
spacy>=3.5.0

evaluate>=0.4.0
